refactor(sandbox): split into per-area files; verbose output; auto-login
- Split src/commands/sandbox.ts into src/commands/sandbox/{index,model,
  lifecycle,auth,inspect}.ts with matching per-area tests. Shared test
  fixtures live in src/testHelpers/sandboxFixtures.ts.
- Every verb now prints step-by-step progress instead of exiting
  silently (ensure/created vs already-exists, regenerate/removing+
  creating+regenerated, rm/removing+removed).
- `auth <model>` drives the agent's login flow directly via the agent's
  own subcommand (claude auth login / codex login / cursor-agent login)
  instead of dropping into the REPL. After exit it runs the matching
  status command and confirms via output pattern. Unknown agents fall
  back to the previous interactive-shell behavior with manual
  verification instructions.
